+++ autofit/non_linear/optimize/pyswarms.py
@@ -12,7 +12,7 @@ from autofit.non_linear.samples import OptimizerSamples, Sample
 from autofit.non_linear.abstract_search import PriorPasser
 from autofit.non_linear.initializer import Initializer
 from autofit.plot import PySwarmsPlotter
-from autofit.plot.mat_wrap.wrap.wrap_base import Output
+from autofit.plot.output import Output
 
 
 class AbstractPySwarms(AbstractOptimizer):
