+++ autofit/plot/__init__.py
@@ -1,24 +1,3 @@
-from autofit.plot.mat_wrap.wrap.wrap_base import Units
-from autofit.plot.mat_wrap.wrap.wrap_base import Figure
-from autofit.plot.mat_wrap.wrap.wrap_base import Axis
-from autofit.plot.mat_wrap.wrap.wrap_base import Cmap
-from autofit.plot.mat_wrap.wrap.wrap_base import Colorbar
-from autofit.plot.mat_wrap.wrap.wrap_base import ColorbarTickParams
-from autofit.plot.mat_wrap.wrap.wrap_base import TickParams
-from autofit.plot.mat_wrap.wrap.wrap_base import YTicks
-from autofit.plot.mat_wrap.wrap.wrap_base import XTicks
-from autofit.plot.mat_wrap.wrap.wrap_base import Title
-from autofit.plot.mat_wrap.wrap.wrap_base import YLabel
-from autofit.plot.mat_wrap.wrap.wrap_base import XLabel
-from autofit.plot.mat_wrap.wrap.wrap_base import Legend
-from autofit.plot.mat_wrap.wrap.wrap_base import Output
-
-from autofit.plot.mat_wrap.mat_plot import MatPlot1D
-from autofit.plot.mat_wrap.include import Include1D
-from autofit.plot.mat_wrap.visuals import Visuals1D
-from autofit.plot.mat_wrap.include import Include2D
-from autofit.plot.mat_wrap.visuals import Visuals2D
-
 from autofit.plot.samples_plotters import SamplesPlotter
 from autofit.plot.dynesty_plotter import DynestyPlotter
 from autofit.plot.ultranest_plotter import UltraNestPlotter
