--- autofit/plot/mat_wrap/wrap/__init__.py
@@ -1,16 +1,0 @@
-from .wrap_base import (
-    Units,
-    Figure,
-    Axis,
-    Cmap,
-    Colorbar,
-    ColorbarTickParams,
-    TickParams,
-    YTicks,
-    XTicks,
-    Title,
-    YLabel,
-    XLabel,
-    Legend,
-    Output,
-)
