+++ autofit/plot/samples_plotters.py
@@ -1,39 +1,23 @@
-from autofit.plot.mat_wrap.wrap import wrap_base
-from autofit.plot.mat_wrap import visuals as vis
-from autofit.plot.mat_wrap import include as inc
-from autofit.plot.mat_wrap import mat_plot as mp
-from autofit.plot import abstract_plotters
-
 import matplotlib.pyplot as plt
+from autofit.plot.output import Output
 
-class SamplesPlotter(abstract_plotters.AbstractPlotter):
+class SamplesPlotter:
     def __init__(
             self, 
             samples,
-            mat_plot_1d: mp.MatPlot1D = mp.MatPlot1D(),
-            visuals_1d: vis.Visuals1D = vis.Visuals1D(),
-            include_1d: inc.Include1D = inc.Include1D(),
-            output : wrap_base.Output = wrap_base.Output()
+            output : Output = Output()
     ):
 
         self.samples = samples
         self.output = output
 
-        super().__init__(
-            mat_plot_1d=mat_plot_1d,
-            visuals_1d=visuals_1d,
-            include_1d=include_1d,
-        )
-
     @property
     def model(self):
         return self.samples.model
 
-    @property
-    def visuals_with_include_2d(self):
-
-        return self.visuals_2d + self.visuals_2d.__class__()
-
+    def close(self):
+        if plt.fignum_exists(num=1):
+            plt.close()
 
 class MCMCPlotter(SamplesPlotter):
 
@@ -52,7 +36,7 @@ class MCMCPlotter(SamplesPlotter):
             ax.set_xlabel(self.model.parameter_labels_latex[i])
 
         self.output.to_figure(structure=None, auto_filename="tracjectories")
-        self.mat_plot_1d.figure.close()
+        self.close()
 
     def _plot_likelihood_series(self, log_posterior_list, **kwargs):
 
@@ -66,7 +50,7 @@ class MCMCPlotter(SamplesPlotter):
         axes.set_xlabel("step number")
 
         self.output.to_figure(structure=None, auto_filename="likelihood_series")
-        self.mat_plot_1d.figure.close()
+        self.close()
 
     def _plot_time_series(self, samples, **kwargs):
 
@@ -80,4 +64,4 @@ class MCMCPlotter(SamplesPlotter):
         axes[-1].set_xlabel("step number")
 
         self.output.to_figure(structure=None, auto_filename="time_series")
-        self.mat_plot_1d.figure.close()
+        self.close()
